fixed maintenace end task for avoiding rake aborted

+++ myapp/lib/tasks/maintenance.rake
@@ -10,6 +10,7 @@ namespace :maintenance do
 
   desc 'end maintenance mode'
   task :end do
+    next unless File.exist?('tmp/maintenace.yml')
     File.delete('tmp/maintenance.yml')
     puts 'off'
   end
